fix:compatible clang on linux

+++ native/scanning.c
@@ -41,7 +41,7 @@ static inline uint64_t add32(uint64_t v1, uint64_t v2, uint64_t *vo) {
 }
 
 static inline uint64_t add64(uint64_t v1, uint64_t v2, uint64_t *vo) {
-    uint64_t v;
+    unsigned long long v;
     uint64_t c = __builtin_uaddll_overflow(v1, v2, &v);
 
     /* set the carry */
